requirements.txt
Django>=4.0
psycopg2>=2.8
gunicorn>=20
djangorestframework
django-cors-headers
django-jsonfield

ansible_runner
ansible-core>==2.12

django-crispy-forms
crispy_bootstrap5
#django-background-tasks

## cron job
#django-crontab

#celery>=5.2.3
#redis